[not verified] Rename component
