+++ projects/plugins/jetpack/extensions/shared/components/product-management-controls/index.js
@@ -7,7 +7,7 @@ import ProductManagementToolbarControl from './toolbar-control';
 
 import './style.scss';
 
-export default function ProductManagementControl( {
+export default function ProductManagementControls( {
 	saveProduct,
 	selectProduct,
 	selectedProductId,
